pyproject.toml
[tool.poetry]
name = "apexagente_cli"
version = "0.1.0"
description = ""
authors = ["jo-project <jo.project.0911@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
jsonwebtoken = "^1.0.0"
colorama = "^0.4.6"
prompt-toolkit = "^3.0.38"
httpx = "^0.24.0"
prettytable = "^3.7.0"
rich = "^13.3.5"
click = "^8.1.3"
termtosvg = "^1.1.0"
fpdf2 = "^2.7.4"
openpyxl = "^3.1.2"
pandas = "^2.0.1"
xlsxwriter = "^3.1.0"

[tool.poetry.scripts]
app = "app.main:main"
pdf = "app.app:main"
test = "app.functions.File:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
